move new service functions to service folder
